feat(axis): band type axis support domain property, closed#162
